# docs/swagger.yaml
basePath: /
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
      trace_id:
        type: object
    type: object
  model.Dept:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Dept'
        type: array
      createBy:
        type: string
      createdAt:
        type: string
      dataScope:
        type: string
      deletedAt:
        type: string
      deptId:
        description: 部门编码
        type: integer
      deptName:
        description: 部门名称
        type: string
      deptPath:
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      params:
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: string
      updateBy:
        type: string
      updatedAt:
        type: string
    type: object
...
host: 127.0.0.1:8060
info:
  contact:
    email: 1@lg1024.com
    name: lg1024
    url: http://lg1024.com
  description: 后台接口
  license: {}
  title: api-gin-web API
  version: 1.0.5
paths:
...
  /api/v1/sys/tables/info:
    post:
      consumes:
      - application/json
      description: 添加表结构
      parameters:
      - description: tableName / 数据表名称
        in: query
        name: tables
        type: string
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 添加表结构
      tags:
      - 工具 - 生成表
    put:
      consumes:
      - application/json
      description: 修改表结构
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Dept'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 修改表结构
      tags:
      - 工具 - 生成表
  /api/v1/sys/tables/info/{tableId}:
    delete:
      description: 删除表结构
      parameters:
      - description: tableId
        in: path
        name: tableId
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "删除失败"}'
          schema:
            type: string
      summary: 删除表结构
      tags:
      - 工具 - 生成表
    get:
      description: 获取JSON
      parameters:
      - description: configKey
        in: path
        name: configKey
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - "": []
      summary: 获取配置
      tags:
      - 工具 - 生成表
  /api/v1/sys/tables/page:
    get:
      description: 生成表分页列表
      parameters:
      - description: tableName / 数据表名称
        in: query
        name: tableName
        type: string
      - description: pageSize / 页条数
        in: query
        name: pageSize
        type: integer
      - description: pageIndex / 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 分页列表数据
      tags:
      - 工具 - 生成表
...
schemes:
- http
- https
securityDefinitions:
  HeaderAuthorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"